+++ Mac/Changelog.markdown
@@ -4,6 +4,7 @@
 
 *Unreleased*
 
+* ⌘S to save the image in place. ⌘E to export the image to a new location.
 * Fix pasting an image when there isn't already an image.